# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api covid19 Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "tagnefabrice@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "KUBEBA API Server"
        }
    ],
    "paths": {
        "/v1/signup": {
            "post": {
                "tags": [
                    "Sign up"
                ],
                "summary": "Register new user",
                "description": "Returns user created",
                "operationId": "Register user",
                "requestBody": {
                    "description": "Pass user information to register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "firt_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "integer",
                                        "example": "Fabrice"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "integer",
                                        "example": "Tagne"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "tagnefabrice@gmail.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "integer",
                                        "example": "656874852"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "account_type": {
                                        "type": "integer",
                                        "example": "client"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/{email}/confirmAccount/{token}": {
            "get": {
                "tags": [
                    "Confirm Account"
                ],
                "summary": "Confirm your account",
                "description": "Returns message",
                "operationId": "Confirm your account",
                "requestBody": {
                    "description": "Pass user email and email verification token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "format": "integer",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "tagnefabrice@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User verified",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/sendEmailVerificationToken/{email}": {
            "get": {
                "tags": [
                    "Confirm Account"
                ],
                "summary": "Send EmailVerification Token",
                "description": "Returns message",
                "operationId": "Send EmailVerification Token",
                "requestBody": {
                    "description": "Pass user email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "tagnefabrice@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User verified",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/forgotpassword": {
            "get": {
                "tags": [
                    "Forgot Password"
                ],
                "summary": "Forgot Password",
                "description": "Returns message",
                "operationId": "Forgot Password",
                "requestBody": {
                    "description": "Pass user email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "tagnefabrice@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verify your email to confirm your account",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/resetPassword": {
            "post": {
                "tags": [
                    "Forgot Password"
                ],
                "summary": "Forgot Password",
                "description": "Returns message",
                "operationId": "Reset Password",
                "requestBody": {
                    "description": "Password revovered",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "tagnefabrice@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset your password to recover your account",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/changePassword": {
            "patch": {
                "tags": [
                    "Change Password"
                ],
                "summary": "Change Password",
                "description": "password modified",
                "operationId": "Change Password",
                "requestBody": {
                    "description": "Change Password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password1"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password1"
                                    },
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Set a new password for your account",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}